# GODRIVER-2914 bson: improve marshal/unmarshal performance by ~58% and ~29%