… bson/bsoncodec/registry_test.go
… bson/testdata/code.json.gz